The resolution of the bounding box (pts as integers) is too low for barcodes. Once the EPSReader can load the HiResBoundingBox this is also useful.
git-svn-id: https://svn.apache.org/repos/asf/xmlgraphics/fop/trunk@370821 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/fop/render/ps/PSImageUtils.java
@@ -198,7 +198,7 @@ public class PSImageUtils {
      */
     public static void renderEPS(byte[] rawEPS, String name,
                     float x, float y, float w, float h,
-                    int bboxx, int bboxy, int bboxw, int bboxh,
+                    float bboxx, float bboxy, float bboxw, float bboxh,
                     PSGenerator gen) throws IOException {
         gen.notifyResourceUsage(PSProcSets.EPS_PROCSET, false);
         gen.writeln("%FOPBeginEPS: " + name);
